Revert "Refactor repacks and add repacked from log view in repack modal"
This reverts commit 8634e5225b00223ecb149448c3c43bcd89932dc8.
