+++ client/packages/common/src/intl/locales/en/inventory.json
@@ -75,7 +75,6 @@
   "messages.no-repacks": "This stock line has not been repacked. Click New to create a new repack.",
   "messages.no-repack-detail": "Select a line to view the details of a repack or click new to create a new repack",
   "messages.all-packs-repacked": "Saved. The previous window has been closed since everything in the stock line has been repacked.",
-  "messages.repack-log-info": "Repacked from stock line",
   "status.new": "New",
   "status.finalised": "Finalised",
   "stocktake.description-template": "Created by {{username}} on {{date}}",
